Clients can chat with other clients
This is the initial implementation of a fully async chat server.
Serialized messages are relayed via IPC from one client to all others.
Receiving clients can deserialize the message and see the text.

TODO: The Message struct properly passes the originating client as
well as a timestamp that the message was originally sent. The sending
client should be able to acknowledge round trip time of a message.